[Refactor] 컴포넌트 분리 중간커밋
