+++ page/friend/FriendInvite.tsx
@@ -2,13 +2,9 @@ import React, {useEffect, useState} from "react";
 import {
     Alert,
     FlatList,
-    Image,
-    KeyboardAvoidingView,
     Modal,
-    Platform,
     StyleSheet,
     Text,
-    TextInput,
     TouchableOpacity,
     View
 } from "react-native";
